requirements.txt
langchain-huggingface
langchain
python-dotenv
langchain_community
langchain_core
streamlit
langchain_groq